UI class refactor, adding fleet lines
fleet lines almost complete
managers are now complete separate from UI, UI extends managers
